change the logic to work

+++ src/utils/users/DeleteUser.tsx
@@ -10,26 +10,25 @@ const DeleteUserPage: NextPage = () => {
     console.log("Handle Submit works");
     const { li: loggedInVal } = router.query;
     console.log("Li: ", loggedInVal);
-    const getuserID = await fetch("/api/tradeEntry/findActID", {
-      method: "POST",
-      body: JSON.stringify(loggedInVal),
-      headers: {
-        "Content-Type": "application/json",
-      },
-    });
-    const userFound = await getuserID.json();
-    console.log("Account ID: ", userFound);
-
-    const done = await fetch("/api/users/deleteUser", {
-      method: "DELETE",
-      body: JSON.stringify({
-        userID: userFound,
-      }),
-    });
-    if (!done.ok) {
-      throw new Error("Failed to delete data");
-    } else if (done.ok) {
-      router.push("/");
+    if (typeof loggedInVal !== "string") {
+      const getuserID = await fetch("/api/tradeEntry/findActID", {
+        method: "POST",
+        body: JSON.stringify(loggedInVal),
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+      const done = await fetch("/api/users/deleteUser", {
+        method: "DELETE",
+        body: JSON.stringify({
+          userID: await getuserID.json(),
+        }),
+      });
+      if (!done.ok) {
+        throw new Error("Failed to delete data");
+      } else if (done.ok) {
+        router.push("/");
+      }
     }
   }
   return (
